Fix decimal issue

+++ src/firebase_cart/models.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from pydantic import BaseModel
 from typing import List, Optional, Dict, Any
 
@@ -10,7 +9,7 @@ class CartItem(BaseModel):
     slug: str
     image: str | None
     quantity: int
-    price: Decimal
+    price: float
 
 
 class Cart(BaseModel):
@@ -21,10 +20,10 @@ class Cart(BaseModel):
     shipping_method: dict
     shipping_address: dict
     billing_address: dict
-    subtotal: Decimal
-    tax_total: Decimal
-    delivery_fee: Decimal
-    total: Decimal
+    subtotal: float
+    tax_total: float
+    delivery_fee: float
+    total: float
     payment_session: dict
 
 
@@ -45,10 +44,10 @@ class Order(BaseModel):
     shipping_method: dict
     shipping_address: dict
     billing_address: dict
-    subtotal: Decimal
-    tax_total: Decimal
-    delivery_fee: Decimal
-    total: Decimal
+    subtotal: float
+    tax_total: float
+    delivery_fee: float
+    total: float
     payment_session: dict
     fulfillments: Optional[list[dict]] = []
     payment_status: Optional[str] = ""
